takes report-inheritance branch and removes changes to school year data modeling
