+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/TargetReportQueryHandler.java
@@ -90,7 +90,7 @@ class TargetReportQueryHandler extends AbstractAggregateQueryHandler<TargetRow,
 
     @Override
     protected TargetReportQuery preProcess(final TargetReportQuery query) {
-        checkArgument(!query.isFilteredSubgroupQuery(), "invalid query, filtered subgroups are not supported");
+        checkArgument(!query.isFilteredSubgroupQuery(), "filtered subgroups are not supported");
         checkArgument(query.getSubjectCode() != null, "invalid query, missing a subject");
         checkArgument(query.getAssessmentGradeCodes().size() == 1, "invalid query, must have one assessment grade");
         checkArgument(query.getSchoolYears().size() == 1, "invalid query, must have exactly one school year");
