+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/AbstractAggregateReportQuery.java
@@ -20,7 +20,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
     private Set<String> subjectCodes;
     private String assessmentTypeCode;
     private ImmutableSet<String> assessmentGradeCodes;
-    private ImmutableSet<Integer> schoolYears;
     private boolean includeState;
     private boolean includeAllDistricts;
     private boolean includeAllSchoolsOfDistricts;
@@ -58,10 +57,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
 
     public Set<String> getAssessmentGradeCodes() {
         return safeImmutableOf(assessmentGradeCodes);
-    }
-
-    public Set<Integer> getSchoolYears() {
-        return safeImmutableOf(schoolYears);
     }
 
     public boolean isIncludeState() {
@@ -131,7 +126,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
         private Set<String> subjectCodes;
         private String assessmentTypeCode;
         private Set<String> assessmentGradeCodes;
-        private Set<Integer> schoolYears;
         private boolean includeState;
         private boolean includeAllDistricts;
         private boolean includeAllSchoolsOfDistricts;
@@ -153,7 +147,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
             reportQuery.subjectCodes = subjectCodes;
             reportQuery.assessmentTypeCode = assessmentTypeCode;
             reportQuery.assessmentGradeCodes = safeImmutableOf(assessmentGradeCodes);
-            reportQuery.schoolYears = safeImmutableOf(schoolYears);
             reportQuery.includeState = includeState;
             reportQuery.includeAllDistricts = includeAllDistricts;
             reportQuery.includeAllSchoolsOfDistricts = includeAllSchoolsOfDistricts;
@@ -177,7 +170,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
                     .subjectCodes(reportQuery.subjectCodes)
                     .assessmentTypeCode(reportQuery.assessmentTypeCode)
                     .assessmentGradeCodes(reportQuery.assessmentGradeCodes)
-                    .schoolYears(reportQuery.schoolYears)
                     .includeState(reportQuery.includeState)
                     .includeAllDistricts(reportQuery.includeAllDistricts)
                     .includeAllSchoolsOfDistricts(reportQuery.includeAllSchoolsOfDistricts)
@@ -207,11 +199,6 @@ public abstract class AbstractAggregateReportQuery extends AbstractReportQuery i
 
         public B assessmentGradeCodes(final Set<String> assessmentGradeCodes) {
             this.assessmentGradeCodes = assessmentGradeCodes;
-            return (B) this;
-        }
-
-        public B schoolYears(final Set<Integer> schoolYears) {
-            this.schoolYears = schoolYears;
             return (B) this;
         }
