+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/ClaimAggregateReportQuery.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.reporting.common.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.ImmutableSet;
 
 import java.util.Map;
 import java.util.Set;
@@ -13,10 +14,17 @@ import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeI
  */
 @JsonIgnoreProperties(value="subjectCodes", allowGetters = true)
 public class ClaimAggregateReportQuery extends AbstractAggregateReportQuery {
+    private ImmutableSet<Integer> schoolYears;
     private ImmutableMap<String, Set<String>> claimCodesBySubject;
 
     public ReportQueryType getType() {
         return ReportQueryType.Claim;
+    }
+
+    // TODO move schoolYears to base report & enforce validation where single year is required
+    @Override
+    public Set<Integer> getSchoolYears() {
+        return safeImmutableOf(schoolYears);
     }
 
     public ImmutableMap<String, Set<String>> getClaimCodesBySubject() {
@@ -38,19 +46,27 @@ public class ClaimAggregateReportQuery extends AbstractAggregateReportQuery {
     }
 
     public static class Builder extends AbstractAggregateReportQuery.Builder<ClaimAggregateReportQuery, Builder> {
+        private Set<Integer> schoolYears;
         private Map<String, Set<String>> claimCodesBySubject;
 
         @Override
         public ClaimAggregateReportQuery build() {
             final ClaimAggregateReportQuery reportQuery = new ClaimAggregateReportQuery();
             super.build(reportQuery);
+            reportQuery.schoolYears = safeImmutableOf(schoolYears);
             reportQuery.claimCodesBySubject = safeImmutableOf(claimCodesBySubject);
             return reportQuery;
         }
 
         public Builder copy(final ClaimAggregateReportQuery value) {
             return super.copy(value)
+                    .schoolYears(value.schoolYears)
                     .claimCodesBySubject(value.claimCodesBySubject);
+        }
+
+        public Builder schoolYears(final Set<Integer> schoolYears) {
+            this.schoolYears = schoolYears;
+            return this;
         }
 
         public Builder claimCodesBySubject(final Map<String, Set<String>> claimCodesBySubject) {
