+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/CustomAggregateReportQuery.java
@@ -1,13 +1,26 @@
 package org.opentestsystem.rdw.reporting.common.model;
 
+import com.google.common.collect.ImmutableSet;
+
+import java.util.Set;
+
+import static org.opentestsystem.rdw.reporting.common.util.CollectionUtils.safeImmutableOf;
+
 /**
  * This class represents a query for the aggregate report with {@link AggregateRow}s
  */
 public class CustomAggregateReportQuery extends AbstractAggregateReportQuery {
 
+    private ImmutableSet<Integer> schoolYears;
+
     @Override
     public ReportQueryType getType() {
         return ReportQueryType.CustomAggregate;
+    }
+
+    // TODO move schoolYears to base report & enforce validation where single year is required
+    public Set<Integer> getSchoolYears() {
+        return safeImmutableOf(schoolYears);
     }
 
     public Builder copy() {
@@ -21,15 +34,24 @@ public class CustomAggregateReportQuery extends AbstractAggregateReportQuery {
 
     public static class Builder extends AbstractAggregateReportQuery.Builder<CustomAggregateReportQuery, Builder> {
 
+        private Set<Integer> schoolYears;
+
+        public Builder schoolYears(final Set<Integer> schoolYears) {
+            this.schoolYears = schoolYears;
+            return this;
+        }
+
         @Override
         public CustomAggregateReportQuery build() {
             final CustomAggregateReportQuery reportQuery = new CustomAggregateReportQuery();
             super.build(reportQuery);
+            reportQuery.schoolYears = safeImmutableOf(schoolYears);
             return reportQuery;
         }
 
         public Builder copy(final CustomAggregateReportQuery reportQuery) {
-            return super.copy(reportQuery);
+            return super.copy(reportQuery)
+                    .schoolYears(reportQuery.schoolYears);
         }
     }
 }
