+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/TargetReportQuery.java
@@ -9,11 +9,19 @@ import java.util.Set;
  * This class represents a query for the aggregate report with {@link TargetRow}s
  */
 public class TargetReportQuery extends AbstractAggregateReportQuery {
+    private int schoolYear;
     private String subjectCode;
 
     @Override
     public ReportQueryType getType() {
         return ReportQueryType.Target;
+    }
+
+    // TODO is this needed? The UI can run a case statement on query type and probably should
+    @Override
+    @JsonIgnore
+    public Set<Integer> getSchoolYears() {
+        return ImmutableSet.of(schoolYear);
     }
 
     @Override
@@ -22,9 +30,8 @@ public class TargetReportQuery extends AbstractAggregateReportQuery {
         return ImmutableSet.of(subjectCode);
     }
 
-    @JsonIgnore
     public int getSchoolYear() {
-        return super.getSchoolYears().iterator().next();
+        return schoolYear;
     }
 
     public String getSubjectCode() {
@@ -45,27 +52,13 @@ public class TargetReportQuery extends AbstractAggregateReportQuery {
         return new Builder();
     }
 
-
     public static class Builder extends AbstractAggregateReportQuery.Builder<TargetReportQuery, Builder> {
+        private int schoolYear;
         private String subjectCode;
-
-        @Override
-        public TargetReportQuery build() {
-            final TargetReportQuery reportQuery = new TargetReportQuery();
-            super.build(reportQuery);
-            reportQuery.subjectCode = subjectCode;
-            return reportQuery;
-        }
-
-        public Builder copy(final TargetReportQuery reportQuery) {
-            super.copy(reportQuery);
-            this.subjectCode = reportQuery.subjectCode;
-            return this;
-        }
 
         // backward compatible method to avoid code changes
         public Builder schoolYear(final int schoolYear) {
-            super.schoolYears(ImmutableSet.of(schoolYear));
+            this.schoolYear = schoolYear;
             return this;
         }
 
@@ -73,6 +66,22 @@ public class TargetReportQuery extends AbstractAggregateReportQuery {
             this.subjectCode = subjectCode;
             return this;
         }
+
+        @Override
+        public TargetReportQuery build() {
+            final TargetReportQuery reportQuery = new TargetReportQuery();
+            super.build(reportQuery);
+            reportQuery.schoolYear = schoolYear;
+            reportQuery.subjectCode = subjectCode;
+            return reportQuery;
+        }
+
+        public Builder copy(final TargetReportQuery reportQuery) {
+            return super.copy(reportQuery)
+                    .schoolYear(reportQuery.schoolYear)
+                    .subjectCode(reportQuery.subjectCode);
+        }
+
     }
 }
