+++ common/src/main/java/org/opentestsystem/rdw/reporting/common/model/jackson/ReportQueryDeserializer.java
@@ -73,17 +73,7 @@ public class ReportQueryDeserializer extends StdDeserializer<AbstractReportQuery
         // this is an aggregate query
         if (node.has("query")) {
             final ObjectNode queryNode = (ObjectNode) node.remove("query");
-
             final String aggregateType = queryNode.remove("reportType").asText();
-
-            if (aggregateType.equals(ReportQueryType.Target.name())) {
-                // target-specific schoolYear was replaced with single-value schoolYears from base class
-                if (queryNode.has("schoolYear")) {
-                    final int schoolYear = queryNode.remove("schoolYear").asInt();
-                    queryNode.putArray("schoolYears").add(schoolYear);
-                }
-            }
-
             queryNode.remove("filteredSubgroupQuery");
             queryNode.put("type", aggregateType);
             queryNode.put("name", name);
